moved method to separate class to make dependency injection possible
