+++ src/main/java/client/gui/IWelcomeMessageUI.java
@@ -1,0 +1,8 @@
+package client.gui;
+
+import java.awt.Dialog;
+
+public interface IWelcomeMessageUI {
+  void showWelcomeMessage();
+  void setDialogVisible(Dialog dialog, boolean visible);
+}
